Added the option to assign tags to sub meshes VisibilitySet and MaterialOverride customizations are now working (in editor)

+++ src/shared/terrain/page.cpp
@@ -338,7 +338,7 @@ namespace mmo
 		float Page::GetSmoothHeightAt(float x, float y) const
 		{
 			// scale down
-			const float scale = constants::PageSize / static_cast<float>(constants::VerticesPerPage - 1);
+			const float scale = static_cast<float>(constants::PageSize / static_cast<double>(constants::VerticesPerPage - 1));
 			x /= scale;
 			y /= scale;
